refactor: 프로필 수정, 팝업 제보 리팩토링
