+++ src/components/my/data/reportCategories.ts
@@ -1,0 +1,10 @@
+import type { ReportCategory } from "@/components/my/types/reportPopup";
+
+export const REPORT_CATEGORIES: ReportCategory[] = [
+  { id: "art", label: "아트/전시", color: "#FFB6D9" },
+  { id: "food", label: "푸드/음료", color: "#FFD4B8" },
+  { id: "fashion", label: "패션/뷰티", color: "#D4C4FF" },
+  { id: "character", label: "캐릭터", color: "#A3B9FF" },
+  { id: "lifestyle", label: "라이프스타일", color: "#B8F0D9" },
+  { id: "etc", label: "기타", color: "#D9F95F" },
+];
