+++ src/components/my/index.ts
@@ -1,6 +1,6 @@
 export { NotificationSettings } from "@/components/my/notifications/NotificationSettings";
 export { Help } from "@/components/my/help/Help";
 export { Settings } from "@/components/my/Settings";
-export { ProfileEdit } from "@/components/my/ProfileEdit";
-export { ReportPopup } from "@/components/my/ReportPopup";
+export { ProfileEdit } from "@/components/my/profile-edit/ProfileEdit";
+export { ReportPopup } from "@/components/my/report-popup/ReportPopup";
 export { My } from "@/components/my/mypage/My";
