+++ src/components/my/types/reportPopup.ts
@@ -1,0 +1,12 @@
+import type { ViewType } from "@/routes/routes";
+
+export type ReportCategory = {
+  id: string;
+  label: string;
+  color: string;
+};
+
+export type ReportPopupProps = {
+  onNavigate: (view: ViewType) => void;
+  breakpoint: "mobile" | "tablet" | "desktop";
+};
